Add request for number of computer in JPA/Criteria and add config for use it

+++ src/main/java/fr/excilys/databasecomputer/Main/Main.java
@@ -76,7 +76,7 @@ public class Main {
 		validator = ValidatorCLI.getInstance();
 		int reponse = 1;
 		int offset = 0;
-		int nbComputer = computerService.nbComputer();
+		long nbComputer = computerService.nbComputer();
 		int maxPage = page.nbPageMax(nbComputer);
 		do {
 			for (Computer computer : computerService.displayAllComputer(page.getLimite(), offset, "ASC")) {
